Sketch out an alternative to `PQgeCopyData`.
Saves CPU time (and power usage) by skipping the allocation and
deallocation of buffers for each row.  The downside is that the data is
only valid for one iteration, but that's probably all that most users of
COPY OUT need anyway.

The callback-based `PQhandleCopyData` reduces CPU usage by more than half
in my toy benchmark.  Wall-clock time stayed about the same as it was.

The disadvantage is that this breaks with existing libpq API style.  But
when performance matters, that might be worth it.